Documentation generation
Added commands for building and serving API documentation.

An internal React project is used to render the API documentation from
the endpoint configuration files.

A new top level node definition node accepts response and request child
nodes that are used to generate the example requests and responses.

A special "@doc" header can be used to create a markdown capable
documentation section for the endpoint that is displayed in the web app.

The faker package has been moved from the endpoint package into its own
package.

The build process has been modified to now include the react build
process along with the Go build. The react app is injected into the
binary on compilation.